adding a feature, which lets you define values to Latching Potis
this is to make it easyer to implement a spindle and feed speed override in my case. Instead of hustle to convert Pins to values in LinuxCNC Hal the component now does it itself. So only the output hast to be linked to the signal of LinuxCNC.
To be tested...